splite ranked pair dv dicts into files for easy exclusion of unused code
