--- nemo-physical/src/dictionary/dv_converter.rs
@@ -1,9 +1,0 @@
-//! This module defines the trait [DvConverter]
-//! for converting datavalues to string,
-//! and also provides serveral implementations.
-
-use crate::datavalues::{AnyDataValue, DataValue, ValueDomain};
-use std::fmt::Debug;
-
-
-
